test(LottoTest, CustomerTest): 한 파일에 합쳐져 있던 단위 테스트를 두 파일로 분리
