+++ __tests__/LottoTest.js
@@ -1,9 +1,4 @@
 const Lotto = require("../src/Lotto");
-const MissionUtils = require("@woowacourse/mission-utils");
-
-const mockQuestions = (answer) => {
-  MissionUtils.Console.readLine = jest.fn().mockReturnValue(answer);
-};
 
 describe("로또 클래스 테스트", () => {
   test("로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.", () => {
@@ -17,19 +12,5 @@ describe("로또 클래스 테스트", () => {
     expect(() => {
       new Lotto([1, 2, 3, 4, 5, 5]);
     }).toThrow("[ERROR]");
-  });
-});
-
-describe("Feat 1. Customer.buyLotto", () => {
-  test("입력된 금액이 천원으로 나누어 떨어진다.", () => {
-    mockQuestions("8000");
-    const app = new App();
-    expect(customer.buyLotto()).toBe(8);
-  });
-
-  test("입력된 금액이 천원으로 나누어 떨어지지 않으면 예외가 발생한다.", () => {
-    mockQuestions("8001");
-    const customer = new Customer();
-    expect(customer.buyLotto).toThrow("[ERROR]");
   });
 });
